myturn/requirements.txt
djangorestframework
channels
channels_redis
django-cors-headers